moving definition of std::complex<stan::math::var> and including it in stan/math/rev/scal.hpp
