+++ stan/math/rev/scal/fun/std_complex.hpp
@@ -1,10 +1,11 @@
-#ifndef STAN_MATH_REV_CORE_STD_COMPLEX_HPP
-#define STAN_MATH_REV_CORE_STD_COMPLEX_HPP
+#ifndef STAN_MATH_REV_SCAL_FUN_STD_COMPLEX_HPP
+#define STAN_MATH_REV_SCAL_FUN_STD_COMPLEX_HPP
 
 #include <stan/math/rev/core.hpp>
 #include <stan/math/rev/core/std_numeric_limits.hpp>
 #include <stan/math/rev/core/std_isnan.hpp>
 #include <stan/math/rev/core/std_isinf.hpp>
+#include <stan/math/rev/scal/fun/is_uninitialized.hpp>
 #include <complex>
 
 namespace std {
@@ -20,6 +21,13 @@ namespace std {
  *
  * Promotion from primitives to stan::math::var applies and this
  * constructor will handle the primitivies.
+ *
+ * This file is located in the stan/math/rev/scal/fun folder
+ * intentionally.  The implementation depends on templated functions
+ * being template instantiated in the correct order. Although the rest
+ * of the std implementations are in stan/math/rev/core, this one is
+ * different because it requires functions defined for
+ * stan::math::var.
  *
  * @param re real component
  * @param im imaginary component
